Move models to a separate Django app
