+++ vulnerabilities/tests/test_api.py
@@ -26,10 +26,10 @@ import os
 
 from django.test import TestCase
 
+from core.models import Package
 from vulnerabilities.api import PackageSerializer
 from vulnerabilities.data_dump import debian_dump
 from vulnerabilities.data_dump import ubuntu_dump
-from vulnerabilities.models import Package
 from vulnerabilities.scraper import debian
 from vulnerabilities.scraper import ubuntu
